extract lbm types into own module
